reinit

+++ dockerfiles/cli/cmd_init.sh
@@ -14,6 +14,7 @@ cmd_init() {
   # set an initial value for the flag
   FORCE_UPDATE="--no-force"
   AUTO_ACCEPT_LICENSE="false"
+  REINIT="false"
 
   while [ $# -gt 0 ]; do
     case $1 in
@@ -23,6 +24,9 @@ cmd_init() {
       --accept-license)
         AUTO_ACCEPT_LICENSE="true"
         shift ;;
+      --reinit)
+        REINIT="true"
+        shift ;;
       *) error "Unknown parameter: $1" ; return 2 ;;
     esac
   done
@@ -30,8 +34,13 @@ cmd_init() {
   if [ "${FORCE_UPDATE}" == "--no-force" ]; then
     # If codenvy.environment file exists, then fail
     if is_initialized; then
-      info "init" "Already initialized."
-      return 1
+      if [[ "${REINIT}" = "true" ]]; then
+        # Save the environment file.
+        cp -rf "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}" "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}".reinit
+      else       
+        info "init" "Already initialized."
+        return 2
+      fi
     fi
   fi
 
@@ -83,31 +92,37 @@ cmd_init() {
     docker_run -v "${CODENVY_HOST_CONFIG}":/copy $IMAGE_INIT
   fi
 
-  # After initialization, add codenvy.env with self-discovery.
-  sed -i'.bak' "s|#CODENVY_HOST=.*|CODENVY_HOST=${CODENVY_HOST}|" "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}"
-  info "init" "  CODENVY_HOST=${CODENVY_HOST}"
+  # After initialization, add environment file.
+  if [[ "${REINIT}" = "true" ]]; then
+    # If this is a reinit, move the .reinit file and use it as the environment file
+    mv -rf "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}".reinit "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}"
+  else       
+    # Otherwise, we are using the templated version and making some modifications.
+    sed -i'.bak' "s|#CODENVY_HOST=.*|CODENVY_HOST=${CODENVY_HOST}|" "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}"
+    info "init" "  CODENVY_HOST=${CODENVY_HOST}"
 
-  sed -i'.bak' "s|#CODENVY_VERSION=.*|CODENVY_VERSION=${CODENVY_VERSION}|" "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}"
-  info "init" "  CODENVY_VERSION=${CODENVY_VERSION}"
-  sed -i'.bak' "s|#CODENVY_CONFIG=.*|CODENVY_CONFIG=${CODENVY_HOST_CONFIG}|" "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}"
-  info "init" "  CODENVY_CONFIG=${CODENVY_HOST_CONFIG}"
-  sed -i'.bak' "s|#CODENVY_INSTANCE=.*|CODENVY_INSTANCE=${CODENVY_HOST_INSTANCE}|" "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}"
-  info "init" "  CODENVY_INSTANCE=${CODENVY_HOST_INSTANCE}"
-  sed -i'.bak' "s|#CODENVY_SWARM_NODES=.*|CODENVY_SWARM_NODES=${CODENVY_HOST}:23750|" "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}"
+    sed -i'.bak' "s|#CODENVY_VERSION=.*|CODENVY_VERSION=${CODENVY_VERSION}|" "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}"
+    info "init" "  CODENVY_VERSION=${CODENVY_VERSION}"
+    sed -i'.bak' "s|#CODENVY_CONFIG=.*|CODENVY_CONFIG=${CODENVY_HOST_CONFIG}|" "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}"
+    info "init" "  CODENVY_CONFIG=${CODENVY_HOST_CONFIG}"
+    sed -i'.bak' "s|#CODENVY_INSTANCE=.*|CODENVY_INSTANCE=${CODENVY_HOST_INSTANCE}|" "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}"
+    info "init" "  CODENVY_INSTANCE=${CODENVY_HOST_INSTANCE}"
+    sed -i'.bak' "s|#CODENVY_SWARM_NODES=.*|CODENVY_SWARM_NODES=${CODENVY_HOST}:23750|" "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}"
 
-  if [ "${CODENVY_DEVELOPMENT_MODE}" == "on" ]; then
-    sed -i'.bak' "s|#CODENVY_ENVIRONMENT=.*|CODENVY_ENVIRONMENT=development|" "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}"
-    info "init" "  CODENVY_ENVIRONMENT=development"
-    sed -i'.bak' "s|#CODENVY_DEVELOPMENT_REPO=.*|CODENVY_DEVELOPMENT_REPO=${CODENVY_HOST_DEVELOPMENT_REPO}|" "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}"
-    info "init" "  CODENVY_DEVELOPMENT_REPO=${CODENVY_HOST_DEVELOPMENT_REPO}"
-    sed -i'.bak' "s|#CODENVY_DEVELOPMENT_TOMCAT=.*|CODENVY_DEVELOPMENT_TOMCAT=${CODENVY_DEVELOPMENT_TOMCAT}|" "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}"
-    info "init" "  CODENVY_DEVELOPMENT_TOMCAT=${CODENVY_DEVELOPMENT_TOMCAT}"
-  else
-    sed -i'.bak' "s|#CODENVY_ENVIRONMENT=.*|CODENVY_ENVIRONMENT=production|" "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}"
-    info "init" "  CODENVY_ENVIRONMENT=production"
+    if [ "${CODENVY_DEVELOPMENT_MODE}" == "on" ]; then
+      sed -i'.bak' "s|#CODENVY_ENVIRONMENT=.*|CODENVY_ENVIRONMENT=development|" "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}"
+      info "init" "  CODENVY_ENVIRONMENT=development"
+      sed -i'.bak' "s|#CODENVY_DEVELOPMENT_REPO=.*|CODENVY_DEVELOPMENT_REPO=${CODENVY_HOST_DEVELOPMENT_REPO}|" "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}"
+      info "init" "  CODENVY_DEVELOPMENT_REPO=${CODENVY_HOST_DEVELOPMENT_REPO}"
+      sed -i'.bak' "s|#CODENVY_DEVELOPMENT_TOMCAT=.*|CODENVY_DEVELOPMENT_TOMCAT=${CODENVY_DEVELOPMENT_TOMCAT}|" "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}"
+      info "init" "  CODENVY_DEVELOPMENT_TOMCAT=${CODENVY_DEVELOPMENT_TOMCAT}"
+    else
+      sed -i'.bak' "s|#CODENVY_ENVIRONMENT=.*|CODENVY_ENVIRONMENT=production|" "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}"
+      info "init" "  CODENVY_ENVIRONMENT=production"
+    fi
+
+    rm -rf "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}".bak > /dev/null 2>&1
   fi
-
-  rm -rf "${REFERENCE_CONTAINER_ENVIRONMENT_FILE}".bak > /dev/null 2>&1
 }
 
 require_license() {
